requirements.txt
aioredis==1.3.1
amqp==5.0.9
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.11
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-filter==21.1
django-templated-mail==1.1.1
djangochannelsrestframework==0.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.3
MarkupSafe==2.0.1
msgpack==1.0.3
oauthlib==3.2.0
packaging==21.3
prompt-toolkit==3.0.28
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
python3-openid==3.2.0
pytz==2021.3
redis==4.1.3
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
Twisted==22.1.0
txaio==21.2.1
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.13.3
zope.interface==5.4.0
python-decouple==3.6

